fix: Governance vote page tab
