+++ dapps/dapp-box/src/modules/NavbarEnhance/Governance/index.tsx
@@ -24,7 +24,7 @@ const GovernanceEnhance: React.FC = () => {
             <NavLink
                 id="governance-vote"
                 name="Vote"
-                path={{ pathname: 'governance/vote/reward-interest-rate', search: location.search }}
+                path={{ pathname: 'governance/vote/proposals', search: location.search }}
                 isCurrent={location.pathname.startsWith('/governance/vote')}
             />
         </div>
